[backend] Ajout script pour envoi mail de rappel
