+++ backend/core/utils.py
@@ -1,7 +1,8 @@
 import csv
 import io
 
-from flask import Response
+from flask import Response, current_app
+from flask_mail import Message
 from werkzeug.datastructures import Headers
 
 
@@ -39,3 +40,58 @@ def transform_obj_to_flat_list(fields, data):
             exp_res[field] = tmp_res
         flat_data.append(exp_res)
     return flat_data
+
+
+def get_mail_subject(text):
+    return current_app.config["ORGANISM_FOR_EMAIL_SUBJECT"] + " " + text
+
+
+def send_email(subject, recipients, html):
+    msg = Message(subject=subject, recipients=recipients, html=html)
+    print("email sent", msg)
+    from app import mail
+    mail.send(msg)
+
+
+def _get_property_names(model_object):
+    from sqlalchemy.ext.hybrid import hybrid_property
+    return [
+        p for p in dir(model_object)
+        if not p.startswith("_")
+           and (
+                   type(model_object.__class__.__dict__.get(p)) == property
+                   or type(model_object.__class__.__dict__.get(p)) == hybrid_property
+           )
+    ]
+
+
+def _get_orm_attribute_names(model_object):
+    return [a for a in model_object.__dict__.keys() if not a.startswith("_")]
+
+
+def stringify(model_object):
+    """Prend une instance de modèle et retourne un dict avec les attributs de l'ORM, les properties et les
+        hybrid properties en clé. Les valeurs sont stringifiées :
+
+        - None -> "--"
+        - True -> "Oui"
+        - False -> "None"
+    """
+    rv = {}
+
+    def process(attributes):
+        for a in attributes:
+            v = getattr(model_object, a)
+            print(a, v)
+            if v is None:
+                v = "--"
+            elif v is True:
+                v = "Oui"
+            elif v is False:
+                v = "Non"
+            rv[a] = v
+
+    process(_get_orm_attribute_names(model_object))
+    process(_get_property_names(model_object))
+
+    return rv
