+++ backend/templates/rappel_event.html
@@ -1,0 +1,5 @@
+<h1>RDV demain pour {{ event.name }} !</h1>
+
+<p>RDV {{ event.hour }} à {{ event.meeting_point }}</p>
+
+<p>{{ event_info.info_rdv }}</p>
